+++ src/lang-xslt/main/uk/co/reecedunn/intellij/plugin/xslt/psi/impl/xslt/XsltParamPsiImpl.kt
@@ -16,7 +16,11 @@
 package uk.co.reecedunn.intellij.plugin.xslt.psi.impl.xslt
 
 import com.intellij.psi.PsiElement
+import uk.co.reecedunn.intellij.plugin.xpath.intellij.resources.XPathIcons
 import uk.co.reecedunn.intellij.plugin.xslt.ast.xslt.XsltParam
 import uk.co.reecedunn.intellij.plugin.xslt.psi.impl.XsltShadowPsiElement
+import javax.swing.Icon
 
-class XsltParamPsiImpl(element: PsiElement) : XsltShadowPsiElement(element), XsltParam
+class XsltParamPsiImpl(element: PsiElement) : XsltShadowPsiElement(element), XsltParam {
+    override fun getIcon(flags: Int): Icon? = XPathIcons.Nodes.Param
+}